+++ DontGoMart/Utilities/Date+Ext.swift
@@ -8,6 +8,11 @@
 import Foundation
 
 extension Date {
+    
+    var dayDisplayFormat: String { self.formatted(.dateTime.day()) }
+    
+    var weekdayDisplayFormat: String { self.formatted(.dateTime.weekday(.wide)) }
+    
     func getAllDates() -> [Date] {
         let calendar = Calendar.current
         let startDate = calendar.date(from: Calendar.current.dateComponents([.year,.month], from: self))!
@@ -18,29 +23,29 @@ extension Date {
     }
     
     func getMonthDayWeekday() -> (month: String, day: String, weekday: String) {
-            let formatter = DateFormatter()
-            formatter.locale = Locale(identifier: "ko_KR") // 한국어 설정
-            
-            // 월
-            formatter.dateFormat = "M월"
-            let month = formatter.string(from: self)
-            
-            // 일
-            formatter.dateFormat = "d일"
-            let day = formatter.string(from: self)
-            
-            // 요일
-            formatter.dateFormat = "EEEE"
-            let weekday = formatter.string(from: self)
-            
-            return (month, day, weekday)
-        }
-}
-
-func dateToDisplay(year: Int = year, month: Int, day: Int) -> Date {
-    let components = DateComponents(calendar: Calendar.current,
-                                    year: year,
-                                    month: month,
-                                    day: day)
-    return Calendar.current.date(from: components) ?? Date()
+        let formatter = DateFormatter()
+        formatter.locale = Locale(identifier: "ko_KR") // 한국어 설정
+        
+        // 월
+        formatter.dateFormat = "M월"
+        let month = formatter.string(from: self)
+        
+        // 일
+        formatter.dateFormat = "d일"
+        let day = formatter.string(from: self)
+        
+        // 요일
+        formatter.dateFormat = "EEEE"
+        let weekday = formatter.string(from: self)
+        
+        return (month, day, weekday)
+    }
+    
+    func dateToDisplay(year: Int = year, month: Int, day: Int) -> Date {
+        let components = DateComponents(calendar: Calendar.current,
+                                        year: year,
+                                        month: month,
+                                        day: day)
+        return Calendar.current.date(from: components) ?? Date()
+    }
 }
