+++ DontGoMartWidget/Widget/TwoHolidayWidget.swift
@@ -9,19 +9,14 @@ import SwiftUI
 import WidgetKit
 
 struct TwoHolidayEntryView: View {
-    @AppStorage("isNormal", store: UserDefaults(suiteName: Utillity.appGroupId)) var isCostco: Bool = false
-    @AppStorage("selectedBranch", store: UserDefaults(suiteName: Utillity.appGroupId)) var selectedBranch: Int = 0
-    @State private var selectedMartType: MartType = .normal
     
     var entry: TwoHolidayEntry
     var config: MonthConfig
-    var widgetDataMapper: WidgetDataMapper
     let startDate = Date()
 
     init(entry: TwoHolidayEntry) {
         self.entry = entry
         self.config = MonthConfig.determineConfig(from: entry.date)
-        self.widgetDataMapper = WidgetDataMapper()
     }
     
     var body: some View {
@@ -50,8 +45,8 @@ struct TwoHolidayWidget: Widget {
         IntentConfiguration(kind: kind, intent: ConfigurationIntent.self, provider: TwoHoliydayWidgetProvider()) { entry in
             TwoHolidayEntryView(entry: entry)
         }
-        .configurationDisplayName("지점별 휴무위젯")
-        .description("두 개의 지점 휴무일을 확인하는 위젯이에요!")
+        .configurationDisplayName("2개의 휴무일 위젯")
+        .description("다다음 휴무일까지 보여주는 위젯이에요!")
         .supportedFamilies([.systemSmall])
     }
 }
